cyber-scanner/requirements.txt
requests>=2.31.0
python-nmap>=0.7.1
rich>=13.7.0
urllib3>=2.0.0
reportlab>=4.0.0
dnspython>=2.6.0
